clean up and move
